+++ db/migrate/20230220095410_add_role_id_tomailinglist.rb
@@ -1,7 +1,7 @@
 class AddRoleIdTomailinglist < ActiveRecord::Migration[5.2]
   def change
     if ENV['MASTER'] == 'true'
-      add_column :mailing_lists, :role_id, :integer
+      add_column :mailing_lists, :role_id, :bigint
       add_foreign_key :mailing_lists, :roles, column: :role_id, primary_key: :role_id
     else
       puts 'Only application on master skipping'
